Switch to Ansible provisioning
- Requires ansible installed on the host

+++ provision/bin/ca-certificates.sh
@@ -4,6 +4,8 @@ set -o errexit
 set -o nounset
 set -o pipefail
 
-echo ">>> Installing Certificate Authorities"
-# DC/OS was compiled on ubuntu with certs in a different place!
-ln -s /etc/ssl/certs/ca-bundle.crt /etc/ssl/certs/ca-certificates.crt
+if [ ! -e /etc/ssl/certs/ca-certificates.crt ]; then
+  echo ">>> Installing Certificate Authorities"
+  # DC/OS was compiled on ubuntu with certs in a different place!
+  ln -s /etc/ssl/certs/ca-bundle.crt /etc/ssl/certs/ca-certificates.crt
+fi
